Make part of ProductList
