+++ week2/aurender-tycoon/aurender-tycoon-duke/AurenderTycoonDuke/AurenderTycoonDuke.csproj
@@ -33,6 +33,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="MySql.Data, Version=6.9.9.0, Culture=neutral, PublicKeyToken=c5687fc88969c44d, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\..\..\..\..\..\..\..\Program Files (x86)\MySQL\MySQL Connector Net 6.9.9\Assemblies\v4.5\MySql.Data.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -43,8 +47,19 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ChangeAble.cs" />
+    <Compile Include="Client.cs" />
+    <Compile Include="ClientManager.cs" />
+    <Compile Include="Invoice.cs" />
+    <Compile Include="Product.cs" />
+    <Compile Include="ProductData.cs" />
+    <Compile Include="ProductManager.cs" />
+    <Compile Include="ProductValue.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Receipt.cs" />
+    <Compile Include="SalesManager.cs" />
+    <Compile Include="SalesStatisticGenerator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
