+++ week2/aurender-tycoon/aurender-tycoon-duke/AurenderTycoonDuke/ChangeAble.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AurenderTycoonDuke
+{
+    interface ChangeAble
+    {
+        int StringToInt(string value);
+    }
+}
