+++ week2/aurender-tycoon/aurender-tycoon-duke/AurenderTycoonDuke/Client.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AurenderTycoonDuke
+{
+    class Client
+    {
+        public string clientName { get; }
+        public string contactInformation { get; }
+        public string destination { get; }
+        public Client(string clientName, string contactInformation, string destination)
+        {
+            this.clientName = clientName;
+            this.contactInformation = contactInformation;
+            this.destination = destination;
+        }
+    }
+}
