+++ week2/aurender-tycoon/aurender-tycoon-duke/AurenderTycoonDuke/ClientManager.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AurenderTycoonDuke
+{
+    class ClientManager
+    {
+        List<Client> clientList = new List<Client>();
+    }
+}
