+++ week2/aurender-tycoon/aurender-tycoon-duke/AurenderTycoonDuke/SalesStatisticGenerator.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AurenderTycoonDuke
+{
+    class SalesStatisticGenerator
+    {
+        public void ViewReceiptBySale()
+        {
+
+        }
+        public void ViewSalesVolume()
+        {
+
+        }
+        public void ViewSalesVolumeByProduct()
+        {
+
+        }
+        public void MakeCsv()
+        {
+
+        }
+    }
+}
